refactor: extract search component
