+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2023-01-20T09:11:14.026Z\n"
-"PO-Revision-Date: 2023-01-20T09:11:14.026Z\n"
+"POT-Creation-Date: 2023-03-31T10:16:40.943Z\n"
+"PO-Revision-Date: 2023-03-31T10:16:40.943Z\n"
 
 msgid "Choose one or more dates..."
 msgstr "Choose one or more dates..."
@@ -819,106 +819,6 @@ msgstr "Create new {{trackedEntityType}}"
 msgid "Register"
 msgstr "Register"
 
-msgid "Fill in at least {{minAttributesRequiredToSearch}}  attributes to search"
-msgstr "Fill in at least {{minAttributesRequiredToSearch}}  attributes to search"
-
-msgid "Search {{name}}"
-msgstr "Search {{name}}"
-
-msgid "Search by {{name}}"
-msgstr "Search by {{name}}"
-
-msgid "Search by attributes"
-msgstr "Search by attributes"
-
-msgid "Fill in these fields to search{{escape}} {{ searchableAttributes }}"
-msgstr "Fill in these fields to search{{escape}} {{ searchableAttributes }}"
-
-msgid ""
-"Fill in at least {{minAttributesRequiredToSearch}} of these fields to "
-"search{{escape}} {{searchableAttributes}}"
-msgstr ""
-"Fill in at least {{minAttributesRequiredToSearch}} of these fields to "
-"search{{escape}} {{searchableAttributes}}"
-
-msgid "Fill in this field to search{{escape}} {{searchableAttributes}}"
-msgstr "Fill in this field to search{{escape}} {{searchableAttributes}}"
-
-msgid "No results found"
-msgstr "No results found"
-
-msgid ""
-"You can change your search terms and search again to find what you are "
-"looking for."
-msgstr ""
-"You can change your search terms and search again to find what you are "
-"looking for."
-
-msgid "Register a user"
-msgstr "Register a user"
-
-msgid "Back to search"
-msgstr "Back to search"
-
-msgid "An error has occurred"
-msgstr "An error has occurred"
-
-msgid "Too many results"
-msgstr "Too many results"
-
-msgid "Cannot search in all programs"
-msgstr "Cannot search in all programs"
-
-msgid "Back"
-msgstr "Back"
-
-msgid "Search for {{titleText}}"
-msgstr "Search for {{titleText}}"
-
-msgid "Search for"
-msgstr "Search for"
-
-msgid "You can also choose a program from the top bar and search in that program"
-msgstr "You can also choose a program from the top bar and search in that program"
-
-msgid "Choose a type to start searching"
-msgstr "Choose a type to start searching"
-
-msgid "all programs"
-msgstr "all programs"
-
-msgid ""
-"Not finding the results you were looking for? Try to search all programs "
-"that use type "
-msgstr ""
-"Not finding the results you were looking for? Try to search all programs "
-"that use type "
-
-msgid "Search in all programs"
-msgstr "Search in all programs"
-
-msgid "If none of search results match, you can create a new "
-msgstr "If none of search results match, you can create a new "
-
-msgid "Create new"
-msgstr "Create new"
-
-msgid "Saved lists in this program"
-msgstr "Saved lists in this program"
-
-msgid "Saved lists offer quick access to your most used views in a program."
-msgstr "Saved lists offer quick access to your most used views in a program."
-
-msgid ""
-"There are no saved lists in this program yet, create one using the button "
-"below."
-msgstr ""
-"There are no saved lists in this program yet, create one using the button "
-"below."
-
-msgid "Create saved list"
-msgstr "Create saved list"
-
 msgid "events"
 msgstr "events"
 
@@ -968,6 +868,9 @@ msgstr "Organisation unit could not be loaded"
 msgid "Possible duplicates found"
 msgstr "Possible duplicates found"
 
+msgid "No results found"
+msgstr "No results found"
+
 msgid "An error occurred loading possible duplicates"
 msgstr "An error occurred loading possible duplicates"
 
@@ -1006,6 +909,103 @@ msgstr "Show all programs"
 
 msgid "Choose a program"
 msgstr "Choose a program"
+
+msgid "Back"
+msgstr "Back"
+
+msgid "Search for {{titleText}}"
+msgstr "Search for {{titleText}}"
+
+msgid "Search for"
+msgstr "Search for"
+
+msgid "You can also choose a program from the top bar and search in that program"
+msgstr "You can also choose a program from the top bar and search in that program"
+
+msgid "Choose a type to start searching"
+msgstr "Choose a type to start searching"
+
+msgid "Fill in at least {{minAttributesRequiredToSearch}}  attributes to search"
+msgstr "Fill in at least {{minAttributesRequiredToSearch}}  attributes to search"
+
+msgid "Search {{name}}"
+msgstr "Search {{name}}"
+
+msgid "Search by {{name}}"
+msgstr "Search by {{name}}"
+
+msgid "Search by attributes"
+msgstr "Search by attributes"
+
+msgid "all programs"
+msgstr "all programs"
+
+msgid ""
+"Not finding the results you were looking for? Try to search all programs "
+"that use type "
+msgstr ""
+"Not finding the results you were looking for? Try to search all programs "
+"that use type "
+
+msgid "Search in all programs"
+msgstr "Search in all programs"
+
+msgid "If none of search results match, you can create a new "
+msgstr "If none of search results match, you can create a new "
+
+msgid "Create new"
+msgstr "Create new"
+
+msgid "Fill in these fields to search{{escape}} {{ searchableAttributes }}"
+msgstr "Fill in these fields to search{{escape}} {{ searchableAttributes }}"
+
+msgid ""
+"Fill in at least {{minAttributesRequiredToSearch}} of these fields to "
+"search{{escape}} {{searchableAttributes}}"
+msgstr ""
+"Fill in at least {{minAttributesRequiredToSearch}} of these fields to "
+"search{{escape}} {{searchableAttributes}}"
+
+msgid "Fill in this field to search{{escape}} {{searchableAttributes}}"
+msgstr "Fill in this field to search{{escape}} {{searchableAttributes}}"
+
+msgid ""
+"You can change your search terms and search again to find what you are "
+"looking for."
+msgstr ""
+"You can change your search terms and search again to find what you are "
+"looking for."
+
+msgid "Register a user"
+msgstr "Register a user"
+
+msgid "Back to search"
+msgstr "Back to search"
+
+msgid "An error has occurred"
+msgstr "An error has occurred"
+
+msgid "Too many results"
+msgstr "Too many results"
+
+msgid "Cannot search in all programs"
+msgstr "Cannot search in all programs"
+
+msgid "Saved lists in this program"
+msgstr "Saved lists in this program"
+
+msgid "Saved lists offer quick access to your most used views in a program."
+msgstr "Saved lists offer quick access to your most used views in a program."
+
+msgid ""
+"There are no saved lists in this program yet, create one using the button "
+"below."
+msgstr ""
+"There are no saved lists in this program yet, create one using the button "
+"below."
+
+msgid "Create saved list"
+msgstr "Create saved list"
 
 msgid "Results found"
 msgstr "Results found"
