+++ src/core_modules/capture-core/components/CardList/CardListItem.component.js
@@ -8,8 +8,8 @@ import { colors, Tag, IconCheckmark16 } from '@dhis2/ui';
 import type {
     CardDataElementsInformation,
     CardProfileImageElementInformation,
-} from '../Pages/Search/SearchResults/SearchResults.types';
-import { searchScopes } from '../Pages/Search/SearchPage.constants';
+} from '../SearchBox';
+import { searchScopes } from '../SearchBox';
 import { enrollmentTypes } from './CardList.constants';
 import { ListEntry } from './ListEntry.component';
 import { dataElementTypes, getTrackerProgramThrowIfNotFound } from '../../metaData';
