+++ src/core_modules/capture-core/components/Pages/NewRelationship/TeiRelationship/SearchResults/TeiRelationshipSearchResults.component.js
@@ -8,7 +8,7 @@ import { withNavigation } from '../../../../Pagination/withDefaultNavigation';
 import { Button } from '../../../../Buttons/Button.component';
 import { makeAttributesSelector } from './teiRelationshipSearchResults.selectors';
 import { CardList } from '../../../../CardList';
-import type { CurrentSearchTerms } from '../../../Search/SearchForm/SearchForm.types';
+import type { CurrentSearchTerms } from '../../../../SearchBox';
 import { SearchResultsHeader } from '../../../../SearchResultsHeader';
 import { type SearchGroup } from '../../../../../metaData';
 import { ResultsPageSizeContext } from '../../../shared-contexts';
