--- src/core_modules/capture-core/components/Pages/Search/SearchForm/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { SearchForm } from './SearchForm.container';
