--- src/core_modules/capture-core/components/Pages/Search/SearchResults/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { SearchResults } from './SearchResults.container';
