--- src/core_modules/capture-core/components/Pages/Search/hooks/index.js
@@ -1,3 +1,0 @@
-
-export { buildSearchGroup } from './buildSearchGroup';
-export { useSearchOption } from './useSearchOption';
