+++ src/core_modules/capture-core/components/Pages/Search/index.js
@@ -1,3 +1,1 @@
 export { SearchPage } from './SearchPage.container';
-export { SearchPageComponent } from './SearchPage.component';
-export { cleanSearchRelatedData, navigateToNewUserPage, showInitialViewOnSearchPage } from './SearchPage.actions';
