+++ src/core_modules/capture-core/components/SearchBox/SearchBox.constants.js
