+++ src/core_modules/capture-core/components/SearchBox/SearchBox.epics.js
@@ -1,0 +1,17 @@
+// @flow
+import { ofType } from 'redux-observable';
+import { EMPTY } from 'rxjs';
+import { switchMap } from 'rxjs/operators';
+import { searchBoxActionTypes } from './SearchBox.actions';
+import { buildUrlQueryString } from '../../utils/routing';
+
+export const navigateToNewUserPageEpic = (action$: InputObservable, store: ReduxStore, { history }: ApiUtils) =>
+    action$.pipe(
+        ofType(searchBoxActionTypes.NAVIGATE_TO_NEW_USER_PAGE),
+        switchMap(() => {
+            const { currentSelections: { programId, orgUnitId, trackedEntityTypeId } } = store.value;
+            history.push(`/new?${buildUrlQueryString({ programId, orgUnitId, trackedEntityTypeId })}`);
+
+            return EMPTY;
+        }),
+    );
