+++ src/core_modules/capture-core/components/SearchBox/SearchForm/index.js
@@ -1,0 +1,12 @@
+// @flow
+export { SearchForm } from './SearchForm.container';
+export type { CurrentSearchTerms } from './SearchForm.types';
+export {
+    searchViaUniqueIdOnScopeProgramEpic,
+    searchViaUniqueIdOnScopeTrackedEntityTypeEpic,
+    searchViaAttributesOnScopeProgramEpic,
+    searchViaAttributesOnScopeTrackedEntityTypeEpic,
+    startFallbackSearchEpic,
+    fallbackSearchEpic,
+    fallbackPushPageEpic,
+} from './SearchForm.epics';
