+++ src/core_modules/capture-core/components/SearchBox/SearchResults/SearchResults.container.js
@@ -4,9 +4,9 @@ import { connect } from 'react-redux';
 import { compose } from 'redux';
 import { SearchResultsComponent } from './SearchResults.component';
 import type { Props, PropsFromRedux, DispatchersFromRedux, OwnProps } from './SearchResults.types';
-import { navigateToNewUserPage, searchViaAttributesOnScopeTrackedEntityType, searchViaAttributesOnScopeProgram, startFallbackSearch } from '../SearchPage.actions';
-import { getTrackedEntityTypeThrowIfNotFound, getTrackerProgramThrowIfNotFound } from '../../../../metaData/helpers';
-import { searchScopes, PAGINATION } from '../SearchPage.constants';
+import { navigateToNewUserPage, searchViaAttributesOnScopeTrackedEntityType, searchViaAttributesOnScopeProgram, startFallbackSearch } from '../SearchBox.actions';
+import { getTrackedEntityTypeThrowIfNotFound, getTrackerProgramThrowIfNotFound } from '../../../metaData/helpers';
+import { searchScopes, PAGINATION } from '../SearchBox.constants';
 
 const getCurrentScope = (scopeId: string, scopeType: $Keys<typeof searchScopes>) => {
     if (!scopeId) {
@@ -37,7 +37,7 @@ const mapStateToProps = (state: ReduxState): PropsFromRedux => {
             formId: currentFormId,
             currentSearchTerms,
         },
-    } = state.searchPage;
+    } = state.searchDomain;
 
     const scopeEntity = getCurrentScope(currentSearchScopeId, currentSearchScopeType);
