+++ src/core_modules/capture-core/components/SearchBox/SearchResults/SearchResults.types.js
@@ -1,9 +1,9 @@
 // @flow
 import type { CurrentSearchTerms } from '../SearchForm/SearchForm.types';
-import { typeof searchScopes } from '../SearchPage.constants';
-import { typeof dataElementTypes } from '../../../../metaData';
-import type { AvailableSearchOption } from '../SearchPage.types';
-import type { ListItem } from '../../../CardList/CardList.types';
+import { typeof searchScopes } from '../SearchBox.constants';
+import { typeof dataElementTypes } from '../../../metaData';
+import type { AvailableSearchOption } from '../SearchBox.types';
+import type { ListItem } from '../../CardList/CardList.types';
 
 
 export type CardDataElementsInformation = Array<{| id: string, name: string, type: $Values<dataElementTypes> |}>
