+++ src/core_modules/capture-core/components/SearchBox/SearchResults/index.js
@@ -1,0 +1,3 @@
+// @flow
+export { SearchResults } from './SearchResults.container';
+export type { CardDataElementsInformation, CardProfileImageElementInformation } from './SearchResults.types';
