+++ src/core_modules/capture-core/components/SearchBox/SearchStatus/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { SearchStatus } from './SearchStatus';
