+++ src/core_modules/capture-core/components/SearchBox/TemplateSelector/TemplateSelectorChip.component.js
