+++ src/core_modules/capture-core/components/SearchBox/TemplateSelector/index.js
