+++ src/core_modules/capture-core/components/SearchBox/TemplateSelector/useTemplates.js
