+++ src/core_modules/capture-core/components/SearchBox/TemplateSelector/workingListsBase.types.js
