+++ src/core_modules/capture-core/components/SearchBox/hooks/index.js
@@ -1,0 +1,5 @@
+
+export { buildSearchGroup } from './buildSearchGroup';
+export { useSearchOption } from './useSearchOption';
+export { usePreselectedProgram } from './usePreselectedProgram';
+export { useFallbackTriggered } from './useFallbackTriggered';
