+++ src/core_modules/capture-core/components/SearchBox/hooks/useFallbackTriggered.js
@@ -1,0 +1,9 @@
+// @flow
+import { useLocation } from 'react-router-dom';
+
+export const useFallbackTriggered = (): boolean => {
+    const defaultValue = { fallback: false };
+    const { state: { fallback } = defaultValue } = useLocation();
+
+    return fallback;
+};
