+++ src/core_modules/capture-core/components/SearchBox/index.js
@@ -1,0 +1,24 @@
+// @flow
+export {
+    cleanSearchRelatedData,
+    navigateToNewUserPage,
+    showInitialViewOnSearchBox,
+    searchBoxActionTypes,
+} from './SearchBox.actions';
+export { SearchBox } from './SearchBox.container';
+export { searchScopes } from './SearchBox.constants';
+export { navigateToNewUserPageEpic } from './SearchBox.epics';
+
+export type { AvailableSearchOption } from './SearchBox.types';
+export type { CardDataElementsInformation, CardProfileImageElementInformation } from './SearchResults';
+export type { CurrentSearchTerms } from './SearchForm';
+export {
+    searchViaUniqueIdOnScopeProgramEpic,
+    searchViaUniqueIdOnScopeTrackedEntityTypeEpic,
+    searchViaAttributesOnScopeProgramEpic,
+    searchViaAttributesOnScopeTrackedEntityTypeEpic,
+    startFallbackSearchEpic,
+    fallbackSearchEpic,
+    fallbackPushPageEpic,
+} from './SearchForm';
+
